src/lib/gladys_client/pyproject.toml
[project]
name = "gladys-client"
version = "0.1.0"
description = "GLADyS gRPC/DB client library"
authors = [
    { name = "Mike Mulcahy" },
    { name = "Scott Mulcahy" },
]
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60",
    "psycopg2-binary>=2.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gladys_client"]
